feat(refactor): extract gameCreationService (PR-15, Phase 1)
First logic-extraction PR of the Phase 1 server-routes refactor. Moves the fat
POST /api/game body into server/services/gameCreationService.ts (class with
deps defaulted to the storage/db/serverGameData singletons + exported
instance, per the AnalyticsService convention).

- POST /api/game handler is now validate-boundary -> service.createGame ->
  respond; the ZodError->400 / else->500 mapping stays in the route. Response
  shape, status codes, and console log lines are byte-equivalent.
- GET /api/game-state reuses gameCreationService.deriveInitialAccessTiers,
  removing its duplicated inline pickTier. Verified the two pre-existing
  pickTier implementations were logic-identical (same descending-threshold
  sort + `rep >= threshold` find + 'none' fallback; only cosmetic brace
  differences) before deduping.
- Dropped the dead `defaultRoles` array (never persisted) with a TODO note.

Characterization test (added in the previous commit) is green before AND
after this swap. tsc clean; full suite 555/555.
